Convert minutes to markdown
Git repositories should not have binary files wherever possible. This commit
removes the ODT file from the repository and replaces it with a Markdown file.

Markdown is a very simple markup language that Github is able to format. This
makes it ideal for circumstances in which LaTeX is not suitable.